+++ test/integration-serial/serial_deploy_test.rb
@@ -161,6 +161,24 @@ class SerialDeployTest < KubernetesDeploy::IntegrationTest
     wait_for_all_crd_deletion
   end
 
+  def test_custom_resources_predeployed
+    assert_deploy_success(deploy_fixtures("crd-predeploy", subset: %w(mail.yml things.yml widgets.yml)))
+    reset_logger
+    assert_deploy_success(deploy_fixtures("crd-predeploy", subset: %w(mail_cr.yml things_cr.yml widgets_cr.yml)))
+    assert_logs_match_all([
+      %r{Phase 3: Predeploying priority resources},
+      %r{Successfully deployed in \d.\ds: Mail/my-first-mail},
+      %r{Successfully deployed in \d.\ds: Thing/my-first-thing},
+      %r{Phase 4: Deploying all resources},
+      %r{Successfully deployed in \d.\ds: Mail/my-first-mail, Thing/my-first-thing, Widget/my-first-widget},
+    ], in_order: true)
+    refute_logs_match(
+      %r{Successfully deployed in \d.\ds: Widget/my-first-widget},
+    )
+  ensure
+    wait_for_all_crd_deletion
+  end
+
   def test_stage_related_metrics_include_custom_tags_from_namespace
     hello_cloud = FixtureSetAssertions::HelloCloud.new(@namespace)
     kubeclient.patch_namespace(hello_cloud.namespace, metadata: { labels: { foo: 'bar' } })